replay/requirements.txt
httpx==0.27.0
pydantic==2.8.2
pyyaml==6.0.2
pytest==8.2.0
